backend/requirements.txt
tensorflow
numpy
opencv-python
pillow
matplotlib
lime
shap
scikit-image
flask
flask-cors
